Revert "BYTES BYTES BYTES"
This reverts commit 6cb600cac40b38ff866efb305fbd637fcee784cf.

+++ database.go
@@ -27,7 +27,7 @@ func (d *Database) Close() {
 	d.db.Close()
 }
 
-func (d *Database) StoreMirror(downstreamID byte, upstreamID byte) error {
+func (d *Database) StoreMirror(downstreamID int, upstreamID int) error {
 
 	// Store the upstream->downstream id
 	d.db.Update(func(tx *bolt.Tx) error {
@@ -46,7 +46,7 @@ func (d *Database) StoreMirror(downstreamID byte, upstreamID byte) error {
 	return nil
 }
 
-func (d *Database) GetDownstreamID(upstreamID byte) []byte {
+func (d *Database) GetDownstreamID(upstreamID int) []byte {
 	var retval = []byte{0}
 	d.db.View(func(tx *bolt.Tx) error {
 		b := tx.Bucket([]byte("up2down"))
@@ -57,7 +57,7 @@ func (d *Database) GetDownstreamID(upstreamID byte) []byte {
 	return retval
 }
 
-func (d *Database) GetUpstreamID(downstreamID byte) []byte {
+func (d *Database) GetUpstreamID(downstreamID int) []byte {
 	var retval = []byte{0}
 	d.db.View(func(tx *bolt.Tx) error {
 		b := tx.Bucket([]byte("down2up"))
